Visually displaying our written files
In FileUtilities.java,

Created a listFiles() method that will get the files directory, create a
file array called filteredFiles that contains all .jpg files found in
our files directory and return these filteredFiles.

In ImageGridFragment.java,

Create a new file array called filteredFiles which will be equal to the
array returned by FileUtilities' listFiles() method. Then, for each file
in this array, decode the bitmap and display it as an ImageGridItem.
Finally, add the ImageGridItem.