Started database transaction
Updated .gitignore to cater for mememaker.iml.

Updated MemeDatasource.java with two helper methods: open() and close()
- only accessible within same class.
Public create() method takes in a meme object, this will be used later.
The create() object opens the database, starts a transaction and then
we'll do some implementation where the comment is. After the
transaction, we'll mark it as successful, end it and close the
connection to the database.