Retrieving our finished memes
Deleted code forcing the database to create itself - this is done upon
app creation.

read() will later be used to read annotations.

readMemes() will read the memes.
To do this, it will open the database.
Then, we'll create a Cursor object (Cursors allow us to store large
amounts of information). This will be set to the group of information
returned when we run a query off the database.
The query will run on MEMES_TABLE and will return columns
COLUMN_MEME_NAME, _ID and COLUMN_MEME_ASSET (These are stored in a new
String array so that they are one parameter in the query method).
Then, we have all of the nulls :)
These nulls are simply filters we can put on the query (So we can group,
order etc... but, we don't want to do this, we just want to return all
of the records).

Then, create a new ArrayList of Meme objects called meme.
Here, we'll move to the cursor to the first record (Will be skipped if
there's no data).
Then, we'll run a do/while loop. So, while there's still records left to
go, create Meme objects!

Once we've got all of our memes, close the cursor, close the database,
close your front door and return the memes.
(Closing front door is optional).

Oh, and the getIntFromColumnName() and getStringFromColumnName() do what
they say on the tin.

getIntFromColumnName() will get the index of the columnName passed in.
We will then return the integer value stored at this index.

getStringFromColumnName() will get the index of the columnName passed
in.
We will then return the String value stored at this index.