Retrieving annotations
Created method addMemeAnnotations() that takes in our list of memes.

Open the database.

For each meme in the memes array...

Create an ArrayList that will store the MemeAnnotations.
Create a Cursor that will run a raw query (We'll just call an SQL
command). This command will select everything in ANNOTATIONS_TABLE where
the meme ID is equal to the ID of the current meme (So we only get
annotations relating to the current meme). We're not including any
selectionArgs so we can pass in null for the 2nd parameter.

Then, as long as we have annotations, create a new MemeAnnotation using
all of the properties passed in.

Then, after the if statement, add our annotations to our array of
annotations and close the cursor.
Then, after the for loop, close the connection to the database.