Editable memes!
You can now edit your meme's annotations!

In CreateMemeActivity.java, if the meme has an ID, we'll update it.
If it doesn't have an ID (It has not yet been created), create it!
Pass in mCurrentMeme as the parameter (This stores all of the
properties).

MemeDatasource now has an update() method that will open the database
and begin a transaction.
We'll then create our ContentValues object so that we can put objects
in.

After that, we will set (or put) meme.getName() as the value for
COLUMN_MEME_NAME.
We'll then run this into an update() method which reads "In MEMES_TABLE,
update the values in updateMemesValue where the ID matches the current
meme ID" the last parameter is null as it is used for whereArgs of which
we do not have.

After that, we are ready to work on updating the annotations.
Here, we have a for loop so that we update each annotation associated to
the meme.
We will create our ContentValues again and put the new values in.

After that, run a check on each individual annotation to see if it has
already been saved.
If it has, run the update() method which will update ANNOTATIONS_TABLE
with the updates updateAnnotations values where BaseColums._ID is equal
to annotation.getId(). Again, the final parameter is for whereArgs which
are not used.
If the annotation hasn't yet been saved (It is new), insert it as a new
annotation with the insert() method. This method will insert the values
of updateAnnotations to ANNOTATIONS_TABLE. The null in the second
parameter is the nullColumnHack we used earlier so that if one of the
values is empty, a NULL will be passed in to avoid an error.

Finally, after the for loop, we'll set the transaction as successful,
end it and close the database.