Memes can be deleted.
New method in MemeDatasource, delete() will open the database and start
a transaction.

We'll then run the delete() method from the database so that we'll
delete from the database where the COLUMN_FOREIGN_KEY_MEME is the same
as memeId (So we delete the selected meme only).

After that, we'll delete the meme itself (We should delete the
annotations first due to the relation with the foreign key which could
give us an error if we try to delete the MEMES_TABLE which holds the
relation first).
To do this, we'll run another delete() command off the database so that
we delete from MEMES_TABLE where the BaseColumns._ID is equal to that of
memeID (Again, so we only delete the one meme).

We'll then mark the transaction as successful and end it.