Code will now take into account database upgrades
If the user is migrating from db V1 to V2, we need to make sure that
they will have a CREATE_DATE value.
Any memes from the V1 database will have the CREATE_DATE value of 0 (The
lowest value).

Also did a bit of refactoring so Meme.java will deal with setTime() and
getTime() methods.

---

First of all, when an upgrade is run, the onUpgrade() method is run.
This is where we use our ALTER_ADD_CREATE_DATE variable.
Here, added a bugfix for where I missed out a space after ADD COLUMN.
This will make sure that the column is created upon migration.

Next, updated the Meme create object so that where we create a new meme,
we pass 0 as a temporary parameter (We must provide a parameter). The
actual creation date will be applied upon saving.
The Meme constructor itself has also been given a createDate parameter
and mCreateDate will be set to the value of what is passed in here.

In SQLite, when we create an integer column, the value defaults to 0 so,
we don't actually need to set the value, we can leave it untouched. That
is why we never actually set a value for our migrated memes.

Finally, in MemeDatasource.java, we updated the query in read() to also
search for the creation date. We'd then get the integer (0 if it was
migrated) of the CREATE_DATE.
Also updated the create() method to set, get and put the time in the
database for the meme.
Note that this code was NOT added in the update() method as we are
talking about a creation date, not a modification date.